# shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Garment Mobile API",
    "description": "Production-ready REST API for Smart Garment cross-platform mobile app",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/cart": {
      "get": {
        "summary": "Get user shopping cart",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Cart items" } }
      }
    },
    "/cart/add": {
      "post": {
        "summary": "Add item to cart",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Item added" } }
      }
    },
...
  }
}